% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_shiny.R
\name{ct_shiny}
\alias{ct_shiny}
\title{A Shiny Interface to the trade and antitrust Packages}
\usage{
ct_shiny()
}
\description{
Launch a shiny interface to simulate the effects of tariffs and mergers
}
\details{
\code{ct_shiny} launches a shiny interface for the \code{antitrust} and \code{trade} packages.
The shiny interface provides users with the ability to calibrate model parameters and simulate
tariff effects using many of the supply and demand models included in the \code{trade} package. It
also provides users with the ability to calibrate different consumer demand systems and simulate the
effects of mergers under different competitive regimes included in the \code{antitrust} package.
}
\examples{
if(interactive()){ct_shiny()}

}
\author{
Charles Taragin, Paulette Wolak
}
