% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicboxmktCnt}
\alias{indicboxmktCnt}
\title{Data on Number of Monte Carlo Simulations Performed in "Indices" Tab}
\format{A data frame with 35 rows and 3 variables:
\describe{
  \item{Cut_type}{price in US dollars (\$326--\$18,823)}
  \item{Cnt}{weight of the diamond (0.2--5.01)}
  \item{shareOutThresh}{quality of the cut (Fair, Good, Very Good, Premium, Ideal)}
}}
\usage{
indicboxmktCnt
}
\description{
A dataset containing the information necessary to calculate the number of merger
simulations used to generate the plots for the "Indices" tab of "Numerical Simulations"
}
\keyword{datasets}
