% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxdata}
\alias{sumboxdata}
\title{Box Plot Statistics for "Summary" Tab}
\format{A data frame with 180 rows and 10 variables:
\describe{
  \item{Demand}{price in US dollars (\$326--\$18,823)}
  \item{Model}{weight of the diamond (0.2--5.01)}
  \item{Outcome}{quality of the cut (Fair, Good, Very Good, Premium, Ideal)}
  \item{Supply}{diamond colour, from D (best) to J (worst)}
  \item{high_wisk}{a measurement of how clear the diamond is (I1 (worst), SI2,
    SI1, VS2, VS1, VVS2, VVS1, IF (best))}
  \item{low_wisk}{length in mm (0--10.74)}
  \item{pct25}{width in mm (0--58.9)}
  \item{pct50}{depth in mm (0--31.8)}
  \item{pct75}{depth in mm (0--31.8)}
  \item{shareOutThresh}{depth in mm (0--31.8)}
}}
\usage{
sumboxdata
}
\description{
A dataset the summary statistics necessary to make boxplots according to supply, demand,
and percent of outside share for horizontal mergers in  order to examine the
relationship between industry price changes and commmonly used merger indices.
}
\keyword{datasets}
