% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxmktCnt}
\alias{sumboxmktCnt}
\title{Data on Number of Monte Carlo Simulations Performed in "Summary"}
\format{A data frame with 151 rows and 6 variables:
\describe{
  \item{Supply}{price in US dollars (\$326--\$18,823)}
  \item{Demand}{weight of the diamond (0.2--5.01)}
  \item{Outcome}{quality of the cut (Fair, Good, Very Good, Premium, Ideal)}
  \item{Cnt}{diamond colour, from D (best) to J (worst)}
  \item{shareOutThresh}{a measurement of how clear the diamond is (I1 (worst), SI2,
    SI1, VS2, VS1, VVS2, VVS1, IF (best))}
  \item{Model}{length in mm (0--10.74)}
}}
\usage{
sumboxmktCnt
}
\description{
A dataset containing the information necessary to calculate the number of merger
simulations used to generate the plots for the "Summary" tab of "Numerical Simulations"
}
\keyword{datasets}
