% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicboxdata}
\alias{indicboxdata}
\title{Box Plot Statistics for "Indices" Tab}
\format{
A data frame with 2,303 rows and 10 variables
\describe{
  \item{Cut_type}{Firm Count, HHI, Delta HHI, UPP, CMCR, Harm 2nd, Party Gap}
  \item{Cut_value}{axis units depending on Cut_type}
  \item{shareOutThresh}{outside share threshold in percent (20--70)}
  \item{Supply}{pooled, bertrand, cournot, auction}
  \item{Demand}{pooled, log, logit, aids, ces, linear}
  \item{high_wisk}{maximum}
  \item{low_wisk}{minimum}
  \item{pct25}{25th percentile boxplot line}
  \item{pct50}{50th percentile boxplot line}
  \item{pct75}{75th percentile boxplot line}
}
}
\usage{
indicboxdata
}
\description{
A dataset containing the summary statistics necessary to make boxplots according to supply, demand,
and percent of outside share for horizontal mergers. This allows for examination of the
relationship between industry price changes and commonly used merger indices.
}
\references{
Taragin, C., & Loudermilk, M. (2019). Using measures of competitive harm for optimal screening of horizontal mergers. mimeo.\doi{10.13140/RG.2.2.30872.85760/1}.
}
\keyword{datasets}
