% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicboxmktCnt}
\alias{indicboxmktCnt}
\title{Number of Monte Carlo Simulations Performed in "Indices" Tab}
\format{
A data frame with 35 rows and 3 variables
\describe{
  \item{Cut_type}{Firm Count, HHI, Delta HHI, UPP, CMCR, Harm 2nd, Party Gap}
  \item{Cnt}{number of horizontal merger simulations (25,890 -- 184,254)}
  \item{shareOutThresh}{outside share threshold in percent (20--70)}
}
}
\usage{
indicboxmktCnt
}
\description{
A dataset containing the information necessary to calculate the number of merger
simulations used to generate the plots for the "Indices" tab of "Numerical Simulations" for
Horizontal Mergers based on the index of interest.
}
\references{
Taragin, C., & Loudermilk, M. (2019). Using measures of competitive harm for optimal screening of horizontal mergers. mimeo.\doi{10.13140/RG.2.2.30872.85760/1}.
}
\keyword{datasets}
