% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxdata}
\alias{sumboxdata}
\title{Box Plot Statistics for "Summary" Tab for Horizontal Mergers}
\format{
A data frame with 210 rows and 10 variables
\describe{
  \item{Demand}{log, logit, aids, ces, linear}
  \item{Model}{cournot:log, cournot: linear, bertrand:aids, bertrand:logit, bertrand:ces, auction:logit}
  \item{Outcome}{post-Merger index of interest (Industry Price Change (percent), Merging Party Price Change (percent), Consumer Harm (dollars), Producer Benefit (dollars), Net Harm (dollars)}
  \item{Supply}{bertrand, cournot, auction}
  \item{high_wisk}{maximum}
  \item{low_wisk}{minimum}
  \item{pct25}{25th percentile boxplot line}
  \item{pct50}{50th percentile boxplot line}
  \item{pct75}{75th percentile boxplot line}
  \item{shareOutThresh}{outside share threshold in percent (20--70) }
}
}
\usage{
sumboxdata
}
\description{
A dataset containing the summary statistics necessary to make boxplots according to supply, demand,
and percent of outside share for horizontal mergers so as to examine the
distribution of outcomes.
}
\references{
Taragin, C., & Loudermilk, M. (2019). Using measures of competitive harm for optimal screening of horizontal mergers. mimeo.\doi{10.13140/RG.2.2.30872.85760/1}.
}
\keyword{datasets}
