% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxdata}
\alias{sumboxdata}
\title{Box Plot Statistics for "Summary" Tab}
\format{
A data frame with 180 rows and 10 variables
\describe{
  \item{Demand}{log, logit, aids, ces, linear}
  \item{Model}{cournot:log, cournot: linear, bertrand:aids, bertrand:logit, bertrand:ces, auction:logit}
  \item{Outcome}{post-Merger index of interest (Industry Price Change (\%), Merging Party Price Change (\%), Consumer Harm (\$), Producer Benefit (\$), Net Harm (\$)}
  \item{Supply}{bertrand, cournot, auction}
  \item{high_wisk}{maximum}
  \item{low_wisk}{minimum}
  \item{pct25}{25th percentile boxplot line}
  \item{pct50}{50th percentile boxplot line}
  \item{pct75}{75th percentile boxplot line}
  \item{shareOutThresh}{outside share threshold in percent (20--70) }
}
}
\usage{
sumboxdata
}
\description{
A dataset containing the summary statistics necessary to make boxplots according to supply, demand,
and percent of outside share for horizontal mergers so as to examine the
distribution of outcomes.
}
\references{
\href{https://www.researchgate.net/publication/330564982_Using_concentration_measures_for_optimal_screening_of_horizontal_mergers}{Taragin and Loudermilk 2019}
}
\keyword{datasets}
