% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxdata_trade}
\alias{sumboxdata_trade}
\title{Box Plot Statistics for "Summary" Tab for Tariffs}
\format{
A data frame with 162 rows and 10 variables
\describe{
  \item{Demand}{Linear, CES, Logit}
  \item{Model}{Cournot:Linear, Bertrand:CES, Bertrand:Logit, Auction2nd:Logit, Bargaining:Logit, Monopolistic Competition:CES, Monopolistic Competition:Logit}
  \item{Outcome}{Consumer Harm, Domestic Firm Benefit, Foreign Firm Harm, Industry Price Change, Net Domestic Harm, Net Total Harm, Domestic Firm Price Change, Foreign Firm Price Change}
  \item{Supply}{Cournot, Bertrand, Auction2nd, Bargaining, Monopolistic Competition}
  \item{high_wisk}{maximum}
  \item{low_wisk}{minimum}
  \item{pct25}{25th percentile boxplot line}
  \item{pct50}{50th percentile boxplot line}
  \item{pct75}{75th percentile boxplot line}
  \item{tariffThresh}{tariff threshold in percent (10--30)}
}
}
\usage{
sumboxdata_trade
}
\description{
A dataset containing the summary statistics necessary to make boxplots according to supply, demand,
and tariff percentage for tariffs so as to examine the
distribution of outcomes.
}
\references{
\href{https://www.researchgate.net/publication/330564982_Using_measures_of_competitive_harm_for_optimal_screening_of_horizontal_mergers}{Taragin and Loudermilk 2019}
}
\keyword{datasets}
