% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxmktCnt}
\alias{sumboxmktCnt}
\title{Number of Monte Carlo Simulations Performed in "Summary" Tab}
\format{
A data frame with 151 rows and 6 variables
\describe{
  \item{Outcome}{post-Merger indice of interest (Industry Price Change (\%), Merging Party Price Change (\%), Consumer Harm (\$), Producer Benefit (\$), Net Harm (\$)}
  \item{Cnt}{number of horizontal merger simulations}
  \item{shareOutThresh}{outside share threshold in percent (20--70)}
}
}
\usage{
sumboxmktCnt
}
\description{
A dataset containing the information necessary to calculate the number of merger
simulations used to generate the plots for the Summary tab of Numerical Simulations.
}
\references{
\href{https://www.researchgate.net/publication/330564982_Using_concentration_measures_for_optimal_screening_of_horizontal_mergers}{Taragin and Loudermilk 2019}
}
\keyword{datasets}
