% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxmktCnt}
\alias{sumboxmktCnt}
\title{Number of Monte Carlo Simulations Performed in "Summary" Tab for Horizontal Mergers}
\format{
A data frame with 30 rows and 3 variables
\describe{
  \item{Outcome}{post-Merger indice of interest (Industry Price Change (percent), Merging Party Price Change (percent), Consumer Harm (dollars), Producer Benefit (dollars), Net Harm (dollars)}
  \item{Cnt}{number of horizontal merger simulations}
  \item{shareOutThresh}{outside share threshold in percent (20--70)}
}
}
\usage{
sumboxmktCnt
}
\description{
A dataset containing the information necessary to calculate the number of merger
simulations used to generate the plots for the Summary tab of Numerical Simulations for Horizontal Mergers.
}
\references{
Taragin, C., & Loudermilk, M. (2019). Using measures of competitive harm for optimal screening of horizontal mergers. mimeo.\doi{10.13140/RG.2.2.30872.85760/1}.
}
\keyword{datasets}
