% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumboxmktCnt_trade}
\alias{sumboxmktCnt_trade}
\title{Number of Monte Carlo Simulations Performed in "Summary" Tab for Tariffs}
\format{
A data frame with 24 rows and 3 variables
\describe{
  \item{Outcome}{Consumer Harm, Domestic Firm Benefit, Foreign Firm Harm, Industry Price Change, Net Domestic Harm, Net Total Harm, Domestic Firm Price Change, Foreign Firm Price Change}
  \item{Cnt}{number of tariffs simulated}
  \item{tariffThresh}{tariff threshold in percent (10--30)}
}
}
\usage{
sumboxmktCnt_trade
}
\description{
A dataset containing the information necessary to calculate the number of tariffs
used to generate the plots for the Summary tab of Numerical Simulations for Tariffs.
}
\references{
\href{https://www.researchgate.net/publication/330564982_Using_measures_of_competitive_harm_for_optimal_screening_of_horizontal_mergers}{Taragin and Loudermilk 2019}
}
\keyword{datasets}
